+++ src/tp4/unq/mercado/ProductoNormal.java
@@ -1,11 +1,11 @@
 package tp4.unq.mercado;
 
 public class ProductoNormal extends Producto{
-	public ProductoNormal(int stock, float precio) {
+	public ProductoNormal(int stock, double precio) {
 		super(stock,precio);
 	}
 	
-	public float precio() {
+	public double precio() {
 		return precio;
 	}
